requirements.txt
# Core dependencies
hydra-core>=1.3.2
omegaconf>=2.3.0
colorlog>=6.7.0

# PDF and document processing
pdfplumber>=0.10.1
pymupdf>=1.23.8
pytesseract>=0.3.10
extract-msg>=0.45.0
python-docx>=1.1.0
pillow>=10.0.0

# ML and embeddings
sentence-transformers>=2.5.0
numpy>=1.24.0

# Storage
neo4j>=5.13.0
chromadb>=0.4.22
duckdb>=0.9.2

# API and services
fastapi>=0.104.1
uvicorn>=0.23.2
pydantic>=2.4.2
grpcio>=1.58.0
grpcio-tools>=1.58.0
graphql-core>=3.2.3

# Development
pytest>=7.4.0
pylint>=3.0.2
black>=23.10.0